requirements.txt
streamlit
streamlit-pandas-profiling
pandas
plotly
altair
requests
db-dtypes
google-cloud-aiplatform
google-cloud-bigquery
google-auth
python-dotenv
ollama
numpy
seaborn
matplotlib
tabulate
statsmodels


sentence-transformers
transformers
huggingface-hub